Create Sieve of Eratosthenes
I recently learned about the Sieve of Eratosthenes in a Discrete Structures lecture, and it was commented by the lecturer that it can be an interesting programming project to undertake for a beginner, I found it challenging in terms of developing the main algorithm that eliminates the non-primes. I am pleased with the final result. I am hoping in the near future to reuse this (in some way, shape or form) for an RSA encryption / decryption program.